Refactor "spin_system.py" into a package
