+++ chemex/parameters/spin_system/__init__.py
@@ -1,0 +1,3 @@
+from .spin_system import Group, PydanticSpinSystem, SpinSystem
+
+__all__ = ["SpinSystem", "PydanticSpinSystem", "Group"]
